Move local and persistent storage modules inside fs
